Ensure all resistors and capacitors in audio path are spec'd correctly.
Basically just went through and made sure we had 0.1% tolerance on the
resistors, with the exception of 1% on the outout resistors because
percision ones (aka current sense resistors ) are $xx.xx, and made sure
that our capacitors are C0G/NP0.  I switched to film capacitors for the
DC blocking input capacitors on the output stage because values that
large for C0G/NP0 ceramics don't exist on Mouser.

Also added ESD protection to the headphone output and line output,
specifically using the TPD2E007, which is specified for AC signals like
ours, and won't activate with our -/+10V rails.