Some housekeeping tasks to keep up the schematic.
Switched from R_US/R_US-VERTICAL to RESISTOR and RESISTOR-HORIZONTAL,
which brings up in line with our usage of CAPACITOR and
CAPACITOR-HORIZONTAL.  Ultimately, our likelihood of using through-hole
passives is very, very low, so I'm trying to commit to pruning down our
rcl (to eventually be named 'passives') library to the most basic/common
SMD passives footprints and nothing more

Also added part numbers, en masse, for many passives.  This is taking
advantage of our new tooling for finding inconsistencies in the
schematic, such as parts missing a part number or identical parts
(device and value) with different part numbers attached to them.  That
still needs to be put into its own repo, and eventually, we'll run it as
part of an automated test against this repository for a continual
feedback loop.  Until then...