Minor board layout tweaks.
Adjusted how we route the USB signals to better take advantage of the
type B connector, which now lets us avoid a layer transition.  We're
about 37mils difference between the pair lengths which is well within
bounds of the 100ps skew allowed by the USB specification.

Also, I mistakenly forgot to route power into the pigtail between the
main board and the headphone detection daughterboard, which is now
rectified.  We're sending it DVDD (3V3) so that it needs no regulator of
its own.  We'll need to make sure that we do have a lot of bulk
capacitance since the pigtail will be 8-12 inches long, and the
measurement current can be large, and thus high di/dt, with frequent
measurements.