Switch everything to vector fonts and add LEDs for knob illumination.
We added LEDs around the volume control potentiometer so that we could
get an underlighting effect on the volume knob.  The base now has
corresponding holes inside the support structure/cavity that the volume
knob sits around, and since we engineered it so that the knob has a
small gap from the base, the new holes added will allow light to escape
all the way out, and under the volume knob, providing a soft white glow
when powered up.

We also switched all text on the board layout to be vector, to better
represent what the actual board will look like.