Pull enclosure files into the repository.
Initially, I had the enclosure files in a separate repository.  As I get
closer to "cutting" a final revision of the board, though, it makes more
sense to bundle it all in one so that I can tag the commit that
represents the finalized board *and* enclosure, ensuring proper fit,
etc.