Adjust power jack footprint and placement.
After getting the test boards back, the PTH fitment is super loose, and
because of that, it's hard to tell if things are positioned correctly.

I went and back took real measurements of the pins, cross-referenced
with what the datasheet says, and ran the numbers again using the IPC
specification for the tightest fit.

I also adjusted the footprint to better account for the centerline of
the part.  Now that the origin of the part is better situated, I
positioned it such that it should yield the optimal board extension of
1mm, letting it sit flush against the enclosure when installed.