Adjust RCA jack placement.
After measuring with the test boards, we had about 1mm worth of distance
from the front face of the RCA jacks to the edge of the board.

Now the jack is shifted over to sit flush with the board edge, exposing
that extra 1mm outside of the enclosure.  This gets us to a more
comfortable 6mm of exposed jack, for better plug seating.