First rough pass at the slimdown.
I've roughly adjusted the dimensions and shoved everything in there.
The routing isn't half bad, but it's gonna need a lot of tweaks, namely
to get things in a better position, with better trace/space, with
cleaner silkscreen layout.

There's still some angles and airwires and obvious stuff to sort out
but, otherwise, a good first pass.